Added Class Diagram.
Added Class Diagram to exemplify how the data should flow, and how the classes should be laid out. Note: Database being a object that belongs to the class is NOT how the functionality should neccesary be; As well as there being more methods than just the ones listed, but to give an idea, due to my lack of creativity.