Added login framework
Added a login framework with some slightly borrowed code that works in any system that has java swing enabled.
The admin account is "admin" and "1234"; while the user account is "user" and "1234". 
The framework is supposed to launch a different window in the package which would be the active program itself. 
The program then would be opened in the appropiate mode